refactor: Use `Sheet` Component for `EmojiGrid` (#32)
* wip(AppGrid): Use Sheet component for EmojiPanel container

* chore(sheet): Add hideClose prop

* style(AppGrid): Use Sheet for EmojiPanel

* refactor: Rename `AppGrid` -> `EmojiGallery`

* chore(sheet): Added `allowOutsideInteraction` prop

* chore(EmojiGallery): Use `Sheet`'s `allowOutsideInteraction`

This corresponds to the modal prop in Radix's Dialog component, which defaults to true. I encountered an issue where pointer-events: none was applied to document.body because of this default. In EmojiGallery, we need to allow interaction outside the sheet while it's open.

As for renaming the prop, the original name modal doesn't make sense in the context of our sheet component.

* fix(EmojiGallery): Tab navigation not selecting next emoji after `EmojiPanel` close